+++ behavex/runner.py
@@ -131,7 +131,7 @@ def run(args):
                 os.environ['FEATURES_PATH'] = features_path + ',' + paths
     features_path = os.environ.get('FEATURES_PATH')
     if features_path == '' or features_path is None:
-        os.environ['FEATURES_PATH'] = os.path.join(os.getcwd(), 'features')
+        os.environ['FEATURES_PATH'] = os.path.abspath(os.path.join('.', 'features'))
     _set_env_variables(args_parsed)
     set_system_paths()
     cleanup_folders()